refactor: Refactor api service
