+++ src/ui-core/pages/WomensClothing/WomensClothing.tsx
@@ -1,11 +1,10 @@
-import React, { useContext} from "react";
 import Header from "../../components/molecules/Header/Header";
 import ItemCard from "../../components/molecules/ItemCard/ItemCard";
 import SectionLayout from "../../layouts/SectionLayout/SectionLayout";
 
 import { Item } from "../../../types/Item"; 
 import { useNavigate } from "react-router-dom"; 
-import { useWoensProducst } from "../../../services/api.service.rq";
+import { useWoensProducts } from "../../../services/product.service";
 import { useUser } from "../../../contexts/UserContext";
 
 export default function WomensClothing() {
@@ -16,7 +15,7 @@ export default function WomensClothing() {
     navigate("/login");
   }
 
-  const { data, isLoading, error } = useWoensProducst();
+  const { data} = useWoensProducts();
   const items = data; 
 
   return (
